backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
aiohttp==3.9.1
asyncio==3.4.3
pydantic==2.5.0
openai==1.3.0
anthropic==0.7.0
websockets==12.0